Improved Gulpfile
Created a 'build' task that prepares code to be be uploaded to Github
Pages. The task compiles less after replacing path(s) to match server's
folder structure.

Currently, all the files are copied to '_gh-pages' directory which is a
git repository in itself, changes can then be committed and pushed.

The 'dev' task, now opens the webserver url upon running.