Improved overall site design
The site now contains a gradient jumbotron which features
platform-specific download buttons.

Stats, like stars and forks, are fethced from Picard's Github
repo and displyed as badges.

Added some content about features of Picard - the text of which
can be improved.

Most of the Bootstrap's features have been disabled via the less
file and will stay so untill a need arises.

Just like the design, most of the code is copied from:
http://getbootstrap.com/2.3.2/

Updates thought for the next iteration:

1. Onhover download button style
2. 'Other downloads' page
3. The giant MB Picard header messes up on small screen - make
   it responsive
4. Currently, the download buttons are visible first and then
   hidden. A better way might be to hide them by default, and
   show only the required one.