Improved download button and features grid
Rather than having multiple download buttons and then hiding one of
them, the page now just contains one button, the properties of which are
modified according to the OS being used. If JS is disabled, the download
button will point to the 'Downloads' page.

The features grid used to look a bit odd, it has now been fixed.