Added Downloads page & Decoupled the less file
A 'Downloads' page has been added which contains all available
downloads.The download links are listed in a table, so that in future
we can easily add a new row for a new version etc.

The design of the download page has been taken from:
https://www.apachefriends.org/download.html

The 'styles.less' has been broken into multiple files for easier
maintenance.