Created a template system using gulp
This is a really basic template system created using the
gulp-file-incude plugin https://www.npmjs.org/package/gulp-file-include

The files that need to be reused are placed in the templates/ folder and
are then included using @@include() in any html file.

The gulp tasks have been largely modified. Now, a temporary build is
generated in the 'tmp/' folder and the webserver now uses 'tmp/' as its
root.

As soon as a file in the 'static/' folder is changed, the change is
reflected in the 'tmp/' folder and the webpage is reloaded.