Escaped lines and improved parsing
I don't know how I missed it, but all the new lines in the NEWS file
start with a * (used as a bullet point.) That was enough to separate
them from other lines.

For escaping the text I've used a snippet from:
http://css-tricks.com/snippets/javascript/htmlentities-for-javascript/
that escapes, albeit very crudely, some html.

I don't think we need a full blown solution, but if we do, I think this
will suffice: https://github.com/mathiasbynens/he