Fixing issues raised in the comments
The file has been renamed to generate_json.

The regexes have been moved out of the function, so they are not
compiled every time the function is called.

The entries in the JSON have been renamed for better readability.
s/ver/version, s/apiver/api_version, s/desc/description.

The JSON now stores the plugins as a dictionary rather than a list, so
searching for plugins is simple.

All variables are noe camelCased.