Updated the plugins page to conform to the new json
Changes to the JSON were introduced in #6.

Beautify has been removed as that is handled by the generate_json
script.

The only thing worth noting is that the description field is assumed to
contain 'safe' html. This filed actually contains html that is used to
beautify the output in Picard's Plugin details box, but here it can be a
bit dangerous (?)