Zipfiles moved to generate_json and other fixes from comments
Files are now zipped when the json is generated, rather than doing
that on every request.

All plugin folders are zipped irrespective of the number of files
they contain - this has been done so as too keep things simple.

The API endpoints have now been versioned, so endpoints are now
/api/v1/plugins/ and /api/v1/download/

In the API, rathen than writing the json to disk on every request,
it's now done after 50 downloads.