Replace spaces by html non-breaking spaces in the list of authors
This was proposed by Sophist, see discussion at https://github.com/musicbrainz/picard-plugins/pull/7#discussion-diff-23888033

The change was reverted because it was doing formatting on the JSON side.

This patch implements a website-side solution doing the same thing.
The nbsp_list function takes care of escaping (hence the `safe` filter in the template).